remove paragraph + section titles
While the paragraph erased is quite fun, it's also extra confusing and unnecessary. On balance, I think this is better if we remove. 

"We have gathered below the insights of many designers of constitutions past. This guidance is separated into those social aspects that concern the crafting of a digital constitution; that is, how to collect and guide the common attention, as well as the material aspects of the constitution; a suggestion for the eventual shape and form that a constitution ought to possess."